Finally found time to reorganize code
