+++ .eslintrc
@@ -1,0 +1,9 @@
+{
+  "env": {
+    "es6": true
+  },
+  "rules": {
+    "strict": 0
+  },
+  "extends": ["mongodb-js/node"]
+}
